+++ main.py
@@ -1,12 +1,11 @@
 import time
 
 import uvicorn
-from fastapi import FastAPI, Request, HTTPException
+from fastapi import FastAPI, HTTPException, Request
 from starlette.middleware.cors import CORSMiddleware
-from pydantic import BaseModel
 
-
-from app.db.access import add_user, database_connect
+from app.db.access import add_Evento, add_rol, add_ticket, add_user, database_connect
+from app.models.models import Evento, Rol, Ticket, User
 
 app = FastAPI()
 
@@ -51,25 +50,6 @@ async def read_main():
 async def root():
     return {"msg": "hola mi rey"}
 
-class User(BaseModel):
-    name: str
-    email: str
-    password: str
-    rol: str
-
-class Ticket(BaseModel):
-    autor: int
-    contenido: str
-    categoria: str
-    prioridad: int
-
-class Evento(BaseModel):
-    ticketID: int
-    contenido: str
-
-class Rol(BaseModel):
-    userID: int
-    rol: str
 
 @app.post("/add_user")
 async def create_user(user: User):
@@ -82,16 +62,20 @@ async def create_user(user: User):
     except Exception as e:
         raise HTTPException(status_code=500, detail=str(e))
 
+
 @app.post("/add_ticket")
 async def create_ticket(ticket: Ticket):
     try:
-        result = add_ticket(ticket.autor, ticket.contenido, ticket.categoria, ticket.prioridad)
+        result = add_ticket(
+            ticket.autor, ticket.contenido, ticket.categoria, ticket.prioridad
+        )
         if result:
             return {"msg": "Ticket ingresado"}
         else:
             raise HTTPException(status_code=500, detail="No se puede :/")
     except Exception as e:
         raise HTTPException(status_code=500, detail=str(e))
+
 
 @app.post("/add_evento")
 async def create_evento(evento: Evento):
@@ -103,6 +87,7 @@ async def create_evento(evento: Evento):
             raise HTTPException(status_code=500, detail="No se puede :/")
     except Exception as e:
         raise HTTPException(status_code=500, detail=str(e))
+
 
 @app.post("/add_rol")
 async def create_rol(rol: Rol):
@@ -116,7 +101,5 @@ async def create_rol(rol: Rol):
         raise HTTPException(status_code=500, detail=str(e))
 
 
-
 if __name__ == "__main__":
     uvicorn.run(app, port=8000)
-
